refactor: use kebab-case convention for files
