+++ lib/setup/setup-avvio.js
@@ -1,5 +1,5 @@
 const Avvio = require('avvio')
-const override = require('../pluginOverride')
+const override = require('../pluginOverride.js')
 const { kState, kAvvioBoot, kKeepAliveConnections } = require('../symbols.js')
 const { hookRunnerApplication } = require('../hooks.js')
