+++ lib/setup/setup-client-error-handler.js
