+++ lib/setup/setup-inject.js
