+++ lib/setup/setup-ready.js
